requirements.txt
asgiref==3.3.0
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.22.0
dj-database-url==0.5.0
Django==3.1.3
django-cloudinary-storage==0.3.0
django-filter==2.4.0
djangoql==0.14.2
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
Markdown==3.3.3
mysqlclient==2.0.1
Pillow==8.0.1
ply==3.11
protobuf==3.13.0
python-dotenv==0.15.0
python-magic==0.4.18
pytz==2020.4
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
